+++ packages/react-screenshots/package.json
@@ -2,13 +2,13 @@
   "name": "react-screenshots",
   "version": "0.0.1-alpha.0",
   "description": "a picture clipping and graffiti tool by react",
-  "main": "dist/index.js",
+  "main": "dist/screenshots.js",
   "files": [
     "dist/**"
   ],
   "scripts": {
     "prepublishOnly": "npm run build",
-    "dev": "npm run dev:Web & npm run dev:electron",
+    "dev": "npm run dev:web & npm run dev:electron",
     "dev:web": "webpack-dev-server --config build/webpack.dev.web.js",
     "dev:electron": "webpack --config build/webpack.dev.electron.js",
     "build": "npm run build:web & npm run build:electron",
@@ -71,8 +71,10 @@
     "less": "^3.10.3",
     "less-loader": "^5.0.0",
     "mini-css-extract-plugin": "^0.8.0",
+    "optimize-css-assets-webpack-plugin": "^5.0.3",
     "postcss-loader": "^3.0.0",
     "style-loader": "^1.0.1",
+    "terser-webpack-plugin": "^2.3.0",
     "url-loader": "^3.0.0",
     "webpack": "^4.41.2",
     "webpack-cli": "^3.3.10",
