+++ packages/react-screenshots/src/components/Screenshots/SizeColor.js
@@ -4,7 +4,7 @@ import Color from './Color'
 
 export default ({ size, color, onSizeChange, onColorChange }) => {
   return (
-    <div className="screenshot-sizecolor">
+    <div className="screenshots-sizecolor">
       <Size value={size} onChange={onSizeChange} />
       <Color value={color} onChange={onColorChange} />
     </div>
