+++ packages/react-screenshots/src/components/Screenshots/actions/save.js
@@ -1,14 +1,14 @@
 import Action from './action'
 
-export default class Ok extends Action {
-  static title = '确定'
+export default class Save extends Action {
+  static title = '保存'
 
-  static icon = 'screenshot-icon-ok'
+  static icon = 'screenshots-icon-save'
 
   constructor (props) {
     super(props)
-    const { el, setContext, context } = props
-    this.emit('onOk', {
+    const { el, context, setContext } = props
+    this.emit('onSave', {
       viewer: { ...context.viewer },
       dataURL: el.toDataURL('image/png')
     })
