+++ packages/react-screenshots/src/components/Screenshots/icons.less
@@ -1,11 +1,11 @@
 @font-face {
-  font-family: 'screenshot-icons';
+  font-family: 'screenshots-icons';
   src: url('./fonts/iconfont.woff') format('woff'), url('./fonts/iconfont.ttf') format('truetype');
 }
 
-[class^='screenshot-icon-'],
-[class*=' screenshot-icon-'] {
-  font-family: 'screenshot-icons' !important;
+[class^='screenshots-icon-'],
+[class*=' screenshots-icon-'] {
+  font-family: 'screenshots-icons' !important;
   font-style: normal;
   font-weight: normal;
   font-variant: normal;
@@ -15,38 +15,38 @@
   -webkit-font-smoothing: antialiased;
 }
 
-.screenshot-icon-ok:before {
+.screenshots-icon-ok:before {
   content: '\e73a';
 }
 
-.screenshot-icon-cancel:before {
+.screenshots-icon-cancel:before {
   content: '\e6fe';
 }
 
-.screenshot-icon-save:before {
+.screenshots-icon-save:before {
   content: '\e6ed';
 }
 
-.screenshot-icon-undo:before {
+.screenshots-icon-undo:before {
   content: '\eaa1';
 }
 
-.screenshot-icon-brush:before {
+.screenshots-icon-brush:before {
   content: '\eaa0';
 }
 
-.screenshot-icon-text:before {
+.screenshots-icon-text:before {
   content: '\e673';
 }
 
-.screenshot-icon-arrow:before {
+.screenshots-icon-arrow:before {
   content: '\e6d7';
 }
 
-.screenshot-icon-rect:before {
+.screenshots-icon-rect:before {
   content: '\e7f9';
 }
 
-.screenshot-icon-ellipse:before {
+.screenshots-icon-ellipse:before {
   content: '\e790';
 }
