+++ packages/react-screenshots/src/components/Screenshots/screenshots.less
@@ -14,7 +14,7 @@
 @button-size: 26px;
 @sizecolor-size: 24px;
 
-.screenshot {
+.screenshots {
   position: relative;
   &,
   * {
